REFACTOR: move socket to other controller
