+++ chat/presentation/src/main/kotlin/vp/togedo/controller/ChatController.kt
@@ -1,14 +1,9 @@
 package vp.togedo.controller
 
 import io.swagger.v3.oas.annotations.Parameter
-import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.messaging.handler.annotation.DestinationVariable
-import org.springframework.messaging.handler.annotation.MessageMapping
 import org.springframework.web.bind.annotation.*
 import reactor.core.publisher.Flux
-import reactor.core.publisher.Mono
-import vp.togedo.data.dto.MessageReqDto
 import vp.togedo.data.dto.MessageResDto
 import vp.togedo.service.ChatService
 
@@ -26,20 +21,4 @@ class ChatController(
         ResponseEntity.ok(
             chatService.getChatMessages(groupId)
         )
-
-    @MessageMapping("/{groupId}")
-    fun publishChatMessage(
-        @Parameter(hidden = true) @RequestHeader("X-VP-UserId") userId: String,
-        @DestinationVariable groupId: String,
-        @RequestBody messageReqDto: MessageReqDto
-    ): Mono<ResponseEntity<Void>> =
-        chatService.publishMessage(
-            groupId = groupId,
-            userId = userId,
-            message = messageReqDto.message
-        ).then<ResponseEntity<Void>?>(Mono.fromCallable { ResponseEntity(HttpStatus.OK) })
-            .doOnNext {
-                println("user id: $userId")
-            }
-
 }
